requirements-dev.txt
# AvivaMente - Development Requirements
# Dependências adicionais para desenvolvimento
# Para instalar: pip install -r requirements-dev.txt

# Inclui requirements principais
-r requirements.txt

# Debug e desenvolvimento
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
coverage==7.3.0
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0

# Code quality
flake8==6.0.0
black==23.7.0
isort==5.12.0

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Database tools
django-seed==0.3.1

# Performance profiling
django-silk==5.0.3

# Mathematical libraries para análises avançadas
numpy==1.25.2
scipy==1.11.2
pandas==2.1.0
matplotlib==3.7.2

# JSON utilities
jsonschema==4.19.1
ujson==5.8.0
